rename class ConfigFilePersistHandler to ConfigFilePersistentHandler
